actually delete content from cache if it was deleted by the user
This will keep the files on info-beamer.com, so we need to manually delete them
after some time.

However, we don't want to delete them right when the user cliecks delete,
because we had to restore files in the past, so we wanna keep them for a bit.

This fixes #11